correction bug hard

+++ hardware.h
@@ -11,7 +11,7 @@ int doit_attendre();*/
 
 ////////////////////////////////
 void print_column(uint16_t column);
-void select_column(int num_column);
+void select_column(uint16_t num_column);
 ////////////////////////////////
 
 void init_hardware();
